+++ src/onegov.user/onegov/user/__init__.py
@@ -5,8 +5,15 @@ log.addHandler(logging.NullHandler())  # noqa
 from translationstring import TranslationStringFactory
 _ = TranslationStringFactory('onegov.user')  # noqa
 
-from onegov.user.auth import Auth, is_valid_yubikey
+from onegov.user.auth import Auth
 from onegov.user.collection import UserCollection
 from onegov.user.model import User
+from onegov.user.utils import is_valid_yubikey, yubikey_otp_to_serial
 
-__all__ = ['Auth', 'is_valid_yubikey', 'User', 'UserCollection']
+__all__ = [
+    'Auth',
+    'is_valid_yubikey',
+    'User',
+    'UserCollection',
+    'yubikey_otp_to_serial'
+]