+++ src/onegov.user/onegov/user/tests/test_auth.py
@@ -2,7 +2,8 @@ import morepath
 import transaction
 
 from more.itsdangerous import IdentityPolicy
-from onegov.user import Auth, is_valid_yubikey, UserCollection
+from onegov.user import (
+    Auth, is_valid_yubikey, UserCollection, yubikey_otp_to_serial)
 from webtest import TestApp as Client
 from unittest.mock import patch
 from yubico_client import Yubico
@@ -53,6 +54,11 @@ def test_is_valid_yubikey_otp(session):
             expected_yubikey_id='ccccccbcgujh',
             yubikey='ccccccbcgujhingjrdejhgfnuetrgigvejhhgbkugded'
         )
+
+
+def test_yubikey_otp_to_serial():
+    assert yubikey_otp_to_serial(
+        'ccccccdefghdefghdefghdefghdefghdefghdefghklv') == 2311522
 
 
 def test_auth_login_yubikey(session):